Split out source file from E form solver header
